Add files via upload
Customer_ Churn _Analysis
INTRODUCTION 
Customer churn is one of the biggest fears of any industry. From various studies in the past, we know that 
the cost of acquiring a new customer has been far greater than retaining one. Churn or churn rate is 
defined as the percentage of customers who stop subscribing to a service or percentage of employees 
leave a job. Churn has affected industries such as banking, insurance, internet streaming and 
telecommunications to just name a few. Although there are many reasons for customer churn, some of 
the major reasons are service dissatisfaction, costly subscription, and better alternatives. Hence, in this 
paper the problem of churning is addressed and data factors affecting the churn are analyzed for their 
effect on the rate.